FPZMSVM
We developed the  FPMZMSVM  package regarding  the method was given by 
Alipour,  Mohammadi,  and  Safari  (2018)  for solar flare prediction.  This package is 
working in MATLAB  workspace.  The  FPMZMSVM uses the Zernike moments
(ZMs)  of the  active  regions (ARs)  (extract for  line-of-sight magnetograms, 
ultraviolet  (UV), and extreme ultraviolet (EUV)  images)  and  the  support vector 
machine (SVM) classifier. We organized this package for:
(a)   A graphical user  (GU)  tool  is applied  for prediction of the flaring and non-flaring 
ARs  using  the  line-of-sight magnetograms  SDO/HMI,  UV  and EUV  images 
SDO/AIA.
(b)   Reproduce the results published in Alipour, Mohammadi, and Safari (2019).